v0.5.0 use size_t replace of uint32_t/unsigned int

+++ libhls/include/hls-segment.h
@@ -6,7 +6,7 @@ struct hls_segment_t
 	int64_t pts;		// present timestamp (millisecond)
 	int64_t duration;	// segment duration (millisecond)
 
-	uint32_t m3u8seq;	// m3u8 file sequence number (base 0)
+	uint64_t m3u8seq;	// m3u8 file sequence number (base 0)
 	uint8_t discontinue; // EXT-X-DISCONTINUITY flag
 
 	char name[128];
